Fixed a series of errors to do with the SPI interface, improved the object orientated-ness of a couple files so they don’t require many getters/setters if any at all. Removed deprecated code. Implemented a simple PID algorithm - which works on all four motors; but over compensates and doesnt return to balanced values…

+++ drone_class.h
@@ -17,7 +17,7 @@ class Drone {
                 bool isLanded();
                 void setLanded(bool);
                 int getAvgSpeed();
-                void updateMotors(double, double, double, int);
+                void updateMotors(double, double, double, double);
         private:
                 PCA9685 pwm;
                 int avg_speed;
